Ajustes en estructura de controladores, añadido manejador de excepciones genéricas.

+++ Kilometros WebAPI/Models/HttpPost/My_Controller/PhysiquePost.cs
@@ -6,7 +6,7 @@ using System.Web;
 namespace Kilometros_WebAPI.Models.HttpPost.My_Controller {
     public class PhysiquePost {
         public short Age;
-        public int Height;
+        public short Height;
         public int Weight;
         public string Sex;
     }
